styleci fix and changed namespace for Requests
