+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\UpdateUserRolesRequest;
+use App\Http\Requests\Setting\UpdateUserRolesRequest;
 use App\Services\GSuiteUserService;
 use App\User;
