+++ app/Http/Requests/Setting/UpdateUserRolesRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Setting;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class SettingRequest extends FormRequest
+class UpdateUserRolesRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,7 +24,8 @@ class SettingRequest extends FormRequest
     public function rules()
     {
         return [
-            'setting_key' => 'required|array',
+            'roles' => 'nullable',
+            'userID' => 'required|integer',
         ];
     }
 }
